projects/image-classifier/requirements.txt
Flask==3.0.0
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.26.2
python-dotenv==1.0.0
Werkzeug==3.0.1
